+++ mrv2/lib/mrvFl/mrvLanguages.cpp
@@ -86,21 +86,6 @@ namespace
             }
         }
 
-        // if (wcschr(cmd, L' ') != NULL)
-        // {
-        //     size_t len = wcslen(cmd) + 3; // 2 for quotes, 1 for null
-        //     terminator LPWSTR quoted_cmd = (LPWSTR)malloc(len *
-        //     sizeof(wchar_t)); if (quoted_cmd == NULL)
-        //     {
-        //         wprintf(L"Failed to allocate memory for command line\n");
-        //         return EXIT_FAILURE;
-        //     }
-        //     swprintf_s(quoted_cmd, len, L"\"%s\"", cmd);
-
-        //     // Free the memory used by the unquoted command
-        //     argv[0] = quoted_cmd;
-        // }
-
         // Call _wexecv with the command string and arguments in separate
         // parameters
         intptr_t result = _wexecv(cmd, argv);